requirements.txt
beautifulsoup4==4.8.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-bootstrap4==1.1.1
django-datatables-view==1.19.1
django-markdownx==3.0.1
django-widget-tweaks==1.4.5
gunicorn==20.0.4
Markdown==3.1.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dotenv==0.10.4
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
static3==0.7.0
whitenoise==4.1.2
